Be_motion

See merge request s13-final/S13P31A103!53

Former-commit-id: c2656dd4bdbc335cd68248ec4f83c90dc243da91

+++ backend/spring-server/src/main/java/com/heungbuja/song/service/SongGameDataCache.java
@@ -111,7 +111,7 @@ public class SongGameDataCache {
                 .songBeat(songBeat)
                 .lyricsInfo(lyricsInfo)
                 .sectionInfo(sectionInfo)
-                .bpm((int) songBeat.getTempoMap().get(0).getBpm())
+                .bpm(songBeat.getTempoMap().get(0).getBpm())
                 .duration(songBeat.getAudio().getDurationSec())
                 .verse1Timeline(verse1Timeline)
                 .verse2Timelines(verse2Timelines)
